% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_models.R
\name{all_models}
\alias{all_models}
\title{Wrapper function for simulations analyzing given data with all models}
\usage{
all_models(data, alpha = 0.025, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} or \code{datasim_cont()} function}

\item{alpha}{Type I error. Default=0.025}
}
\value{
List containing an indicator whether the null hypothesis was rejected or not for the investigated treatment for all models
}
\description{
Analyzes given data using the fixed effect model, pooled and separate analyses, and timemachine and MAP prior approach. Performs inference for all treatment arms in the trial except for the first one
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

all_models(data = trial_data)


}
\author{
Pavla Krotka
}
\keyword{internal}
