% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_functions.R
\name{calc_fat_mass}
\alias{calc_fat_mass}
\title{Calculate fat mass}
\usage{
calc_fat_mass(bm, ifbfat)
}
\arguments{
\item{bm}{Numeric. Body mass (kg).}

\item{ifbfat}{Numeric. Ingesta-free body fat as a fraction of body mass.}
}
\value{
Numeric. Fat mass (kg).
}
\description{
Returns the fat component of body mass from the ingesta-free
body fat fraction.
}
\keyword{internal}
