% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{Data simulation for a platform trial with non-concurrent controls with an arbitrary number of treatments and periods}
\usage{
data_sim(
  SS_matrix,
  block_sizes,
  alloc_ratios,
  mu0 = 0,
  delta,
  p0,
  OR,
  lambda,
  sigma,
  N_peak,
  trend,
  trend_param,
  endpoint
)
}
\arguments{
\item{SS_matrix}{a}

\item{block_sizes}{a}

\item{alloc_ratios}{a}

\item{mu0}{a}

\item{delta}{a}

\item{p0}{a}

\item{OR}{a}

\item{lambda}{a}

\item{sigma}{a}

\item{N_peak}{a}

\item{trend}{a}

\item{trend_param}{a}

\item{endpoint}{a}
}
\value{

}
\description{

}
\references{

}
\author{

}
