% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim_cont}
\alias{data_sim_cont}
\title{Data simulation for continuous endpoints for a platform trial with 3 treatment arms entering sequentially}
\usage{
data_sim_cont(
  T_,
  N,
  SS_arm,
  alloc_ratios,
  block_sizes,
  mu0 = 0,
  delta,
  lambda,
  sigma,
  N_peak,
  trend
)
}
\arguments{
\item{T_}{Timing of adding new arms in terms of number of patients allocated to the control arm}

\item{N}{Overall sample size for the trial}

\item{SS_arm}{Sample size per arm}

\item{alloc_ratios}{Matrix with allocation ratios in each period}

\item{block_sizes}{Vector with block size in each period}

\item{mu0}{Response in the control arm. Default=0}

\item{delta}{Vector with treatment effects for each arm}

\item{lambda}{Vector with strength of time trend in each arm}

\item{sigma}{Residual variance}

\item{N_peak}{Point at which the inverted-u time trend switches direction in terms of overall sample size}

\item{trend}{Indicates the time trend pattern ("linear", "stepwise" or "inv_u")}
}
\value{
Data frame: simulated trial data
}
\description{
Simulates data from a platform trial with 3 treatment arms entering sequentially. There is the option to either specify the timing of adding arms and the overall sample size in the trial or specify the sample size per arm (assumed equal) and the allocation ratios per period.
}
\author{
Pavla Krotka, Marta Bofill Roig
}
