% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim_cont.R
\name{datasim_cont}
\alias{datasim_cont}
\title{Data simulation for continuous endpoints for a platform trial with an arbitrary number of treatment arms entering sequentially}
\usage{
datasim_cont(
  num_arms,
  n_arm,
  d,
  period_blocks = 2,
  mu0 = 0,
  theta,
  lambda,
  sigma,
  trend,
  N_peak,
  full = FALSE
)
}
\arguments{
\item{num_arms}{Number of treatment arms in the trial}

\item{n_arm}{Sample size per arm (assumed equal)}

\item{d}{Vector with timings of adding new arms in terms of number of patients recruited to the trial so far (of length num_arms)}

\item{period_blocks}{Number to multiply the number of active arms with in order to get the block size per period}

\item{mu0}{Response in the control arm. Default=0}

\item{theta}{Vector with treatment effects for each treatment arm (of length num_arms)}

\item{lambda}{Vector with strength of time trend in each arm (of length num_arms+1, as time trend in the control is also allowed)}

\item{sigma}{Residual variance}

\item{trend}{Indicates the time trend pattern ("linear", "stepwise", "stepwise_2", or "inv_u")}

\item{N_peak}{Point at which the inverted-u time trend switches direction in terms of overall sample size}

\item{full}{Boolean. Indicates whether only variables needed for the analysis should be in the output (FALSE) or also additional information (lambdas, underlying responses, input parameters) should be included (TRUE). Default=FALSE}
}
\value{
Data frame: simulated trial data (Default, if full=FALSE) or List: simulated trial data, input parameters and additional information (if full=TRUE)
}
\description{
Simulates data from a platform trial with binary endpoints and an arbitrary number of treatment arms entering sequentially. The user specifies the timing of adding arms in terms of patients recruited to the trial so far and the the sample size per arm (assumed equal).
}
\examples{

head(datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear"))


}
\author{
Pavla Krotka, Marta Bofill Roig
}
