% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmodel_bin.R
\name{fixmodel_bin}
\alias{fixmodel_bin}
\title{Model-based analysis for binary data adjusting for periods}
\usage{
fixmodel_bin(data, arm, alpha = 0.025)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Performs logistic regression taking into account all trial data until the arm under study leaves the trial and adjusting for periods as factors
}
\examples{

trial_data <- datasim_bin(n_total = 1000, num_arms = 3, d = 120,
p0 = 0.7, OR = rep(1.4, 3), lambda = rep(0.15, 4), trend = "linear")

fixmodel_bin(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
