% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmodel_cal_bin.R
\name{fixmodel_cal_bin}
\alias{fixmodel_cal_bin}
\title{Frequentist logistic regression model analysis for binary data adjusting for calendar time units}
\usage{
fixmodel_cal_bin(
  data,
  arm,
  alpha = 0.025,
  unit_size = 25,
  ncc = TRUE,
  check = TRUE,
  ...
)
}
\arguments{
\item{data}{Trial data, e.g. result from the \code{datasim_bin()} function. Must contain columns named 'treatment' and 'response'.}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.}

\item{alpha}{Significance level. Default=0.025.}

\item{unit_size}{Number of patients per calendar time unit. Default=25.}

\item{ncc}{Boolean. Whether to include NCC data into the analysis. Default=TRUE.}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.}

\item{...}{Further arguments for simulation function.}
}
\value{
List containing the following elements regarding the results of comparing \code{arm} to control:
\itemize{
\item \code{p-val} - p-value (one-sided)
\item \code{treat_effect} - estimated treatment effect in terms of the log-odds ratio
\item \code{lower_ci} - lower limit of the 95\% confidence interval
\item \code{upper_ci} - upper limit of the 95\% confidence interval
\item \code{reject_h0} - indicator of whether the null hypothesis was rejected or not (\code{p_val} < \code{alpha})
\item \code{model} - fitted model
}
}
\description{
This function performs logistic regression taking into account all trial data until the arm under study leaves the trial and adjusting for calendar time units as factors.
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

fixmodel_cal_bin(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
