% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmodel_cont.R
\name{fixmodel_cont}
\alias{fixmodel_cont}
\title{Model-based analysis for continuous data adjusting for periods}
\usage{
fixmodel_cont(data, arm, alpha = 0.025)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Performs linear regression taking into account all trial data until the arm under study leaves the trial and adjusting for periods as factors
}
\examples{

trial_data <- datasim_cont(n_total = 1000, num_arms = 3, d = 120,
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

fixmodel_cont(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
