% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_cont.R
\name{gam_cont}
\alias{gam_cont}
\title{Model-based analysis for continuous data using generalized additive model}
\usage{
gam_cont(
  data,
  arm,
  alpha = 0.025,
  ci = FALSE,
  smoothing_basis = "tp",
  basis_dim = -1,
  gam_method = "GCV.Cp",
  ...
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{ci}{Boolean. Whether confidence intervals should be computed. Default=FALSE}

\item{smoothing_basis}{String indicating the (penalized) smoothing basis to use. Default="tp"}

\item{basis_dim}{The dimension of the basis used to represent the smooth term. The default depends on the number of variables that the smooth is a function of.}

\item{gam_method}{The smoothing parameter estimation method. Default="GCV.Cp"}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, an indicator whether the null hypothesis was rejected or not for the investigated treatment and the fitted model
}
\description{
Performs analysis using a generalized additive model taking into account all trial data until the arm under study leaves the trial
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

gam_cont(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
