% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alloc_matrix.R
\name{get_alloc_matrix}
\alias{get_alloc_matrix}
\title{Sample size matrix for a platform trial with 3 arms}
\usage{
get_alloc_matrix(n_total, n_arm, num_arms, d, alloc_ratio = c(1, 1))
}
\arguments{
\item{n_total}{Overall sample size for the trial}

\item{n_arm}{Sample size per arm}

\item{num_arms}{Number of treatment arms in the trial}

\item{d}{Timing of adding new arms in terms of number of patients allocated to the control arm}

\item{alloc_ratio}{Vector of lenght 2. Allocation ratio control vs treatment}
}
\value{
Sample size matrix
}
\description{
Computes the sample size matrix - sample sizes per arm (rows) and per period (columns)
}
\author{
Pavla Krotka
}
\keyword{internal}
