% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss_matrix.R
\name{get_ss_matrix}
\alias{get_ss_matrix}
\title{Sample size matrix for a platform trial with 3 arms}
\usage{
get_ss_matrix(T_, N)
}
\arguments{
\item{T_}{Timing of adding new arms in terms of number of patients allocated to the control arm}

\item{N}{Overall sample size for the trial}
}
\value{
Sample size matrix
}
\description{
Computes the sample size matrix - sample sizes per arm (rows) and per period (columns)
}
\author{
Pavla Krotka
}
\keyword{internal}
