% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_trend2.R
\name{linear_trend2}
\alias{linear_trend2}
\title{Generation of a linear trend that starts in the second period}
\usage{
linear_trend2(j, lambda, sample_size)
}
\arguments{
\item{j}{a}

\item{lambda}{a}

\item{sample_size}{vector of dimension 2, indicating sample size in the first period and the remaining sample size}
}
\value{

}
\description{

}
\references{

}
\author{

}
