% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_rjags_cont.R
\name{MAP_rjags_cont}
\alias{MAP_rjags_cont}
\title{Remarks'}
\usage{
MAP_rjags_cont(
  data,
  arm,
  alpha = 0.025,
  opt = 2,
  prior_prec_tau = 4,
  n.samples = 1000,
  n.chains = 4,
  n.iter = 4000,
  n.adapt = 1000,
  robustify = TRUE,
  weight = 0.1,
  ...
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{opt}{binary, opt=1: all former periods are used as one source, opt=2: periods form different sources get seperately included into the final analysis}

\item{prior_prec_tau}{dispersion parameter of the half normal prior, the prior for the between study heteroginity}

\item{n.samples}{defines the number of how many random samples will get drwan for the calcualtion of the posterior mean and the CIs, default is 1000}

\item{n.chains}{the number of parallel chains for the rjags model}

\item{n.iter}{needed for coda.samples, number of iterations to monitor of the jags.model}

\item{n.adapt}{needed for jags.model, defines the number of iterations for adaptation, an initial sampling phase during which the samplers adapt their behaviour to maximize their efficiency}

\item{robustify}{Boolean.}

\item{weight}{weight given to the non-informative component (0 < weight < 1) for the robustification of the MAP prior according to Schmidli 2014}

\item{...}{Further arguments for simulation function}
}
\value{
containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Borrows data from non-concurrent controls for binary endpoints for the concurrent controls arm by the use of a MAP prior
}
\details{
Check input parmaters for prior specification and whether to include additional ones like robistify yes, no, specifications for robustification in the future...
So far: Posterior odds estimated by drawing random samples from the posterior distributions of the incidence rates of the two groups seperately
}
\examples{

trial_data<-datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "stepwise")

MAP_rjags_cont(data = trial_data, arm = 3)


}
\author{
Katharina Hees
}
