% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPprior_bin.R
\name{MAPprior_bin}
\alias{MAPprior_bin}
\title{To-DOs'}
\usage{
MAPprior_bin(
  data,
  arm,
  alpha = 0.025,
  opt = 1,
  tau.dist = "HalfNormal",
  tau.prior = 1,
  beta.prior = 2,
  prec.ef = 1000
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{opt}{Binary, opt=1: all former periods are use as one source, opt=2: periods form different sources get separately included into the final analysis}

\item{tau.dist}{Type of prior distribution for tau; supported priors are HalfNormal (default), TruncNormal, Uniform, Gamma, InvGamma, LogNormal, TruncCauchy, Exp and Fixed, default is "HalfNormal", for more details see help of the gMAP function of the RBesT package}

\item{tau.prior}{Parameters of prior distribution for tau, default is 1}

\item{beta.prior}{Mean and standard deviation for normal priors of regression coefficients, see section prior specification help of the gMAP function of the RBesT package, default is 2
tau.dist="HalfNormal",tau.prior=1,beta.prior=2}

\item{prec.ef}{Defines the number of how many random samples will get drawn for the calculation of the posterior mean and the CIs, default is 1000}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Borrows data from non-concurrent controls for binary endpoints for the concurrent controls arm by the use of a MAP prior
}
\details{
Check input parameters for prior specification and whether to include additional ones like robistify yes, no, specifications for robustification,...
Check p-value and different definitions of it : posterior p-value, posterior predictive p-value, posterior probability,...
Posterior odds estimated by drawing random samples from the posterior distributions of the incidence rates of the two groups separately

Check: Take into account the following warning statement?: "In total 1 divergent transitions occurred during the sampling phase.
Please consider increasing adapt_delta closer to 1 with the following command prior to gMAP:
options(RBesT.MC.control=list(adapt_delta=0.999))" ?

Use of non-concurrent controls for binary data
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

MAPprior_bin(data = trial_data, arm = 3)


}
\author{
Katharina Hees
}
