% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPprior_cont.R
\name{MAPprior_cont}
\alias{MAPprior_cont}
\title{Analysis for continuous data using the MAP Prior approach}
\usage{
MAPprior_cont(
  data,
  arm,
  alpha = 0.025,
  opt = 2,
  prior_prec_tau = 4,
  prior_prec_eta = 0.001,
  n.samples = 1000,
  n.chains = 4,
  n.iter = 4000,
  n.adapt = 1000,
  robustify = TRUE,
  weight = 0.1,
  check = TRUE,
  ...
)
}
\arguments{
\item{data}{Trial data, e.g. result from the \code{datasim_bin()} function. Must contain columns named 'treatment', 'response' and 'period'.}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.}

\item{alpha}{Significance level. Default=0.025}

\item{opt}{Binary. If opt==1, all former periods are used as one source; if opt==2, periods get separately included into the final analysis. Default=2.}

\item{prior_prec_tau}{Dispersion parameter of the half normal prior, the prior for the between study heterogeneity. Default=4.}

\item{prior_prec_eta}{Dispersion parameter of the normal prior, the prior for the control mean. Default=0.001.}

\item{n.samples}{Number of how many random samples will get drawn for the calculation of the posterior mean, the p-value and the CI's. Default=1000.}

\item{n.chains}{Number of parallel chains for the rjags model. Default=4.}

\item{n.iter}{Number of iterations to monitor of the jags.model. Needed for coda.samples. Default=4000.}

\item{n.adapt}{Number of iterations for adaptation, an initial sampling phase during which the samplers adapt their behavior to maximize their efficiency. Needed for jags.model. Default=1000.}

\item{robustify}{Boolean. Indicates whether a robust prior is to be used. If TRUE, a mixture prior is considered combining a MAP prior and a weakly non-informative component prior. Default=TRUE.}

\item{weight}{Weight given to the non-informative component (0 < weight < 1) for the robustification of the MAP prior according to Schmidli (2014). Default=0.1.}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.}

\item{...}{Further arguments for simulation function.}
}
\value{
List containing the following elements regarding the results of comparing \code{arm} to control:
\itemize{
\item \code{p-val} - p-value (one-sided) obtained by drawing \code{n.samples} random samples from each posterior distribution
\item \code{treat_effect} - estimated treatment effect in terms of the difference in means obtained by drawing \code{n.samples} random samples from each posterior distribution
\item \code{lower_ci} - lower limit of the 95\% confidence interval obtained by drawing \code{n.samples} random samples from each posterior distribution
\item \code{upper_ci} - upper limit of the 95\% confidence interval obtained by drawing \code{n.samples} random samples from each posterior distribution
\item \code{reject_h0} - indicator of whether the null hypothesis was rejected or not (\code{p_val} < \code{alpha})
}
}
\description{
This function performs analysis of continuous data using the Meta-Analytic-Predictive (MAP) Prior approach. The method borrows data from non-concurrent controls to obtain the prior distribution for the control response in the concurrent periods.
}
\details{
In this function, the argument \code{alpha} corresponds to \eqn{1-\gamma}, where \eqn{\gamma} is the decision boundary. Specifically, the posterior probability of the difference distribution under the null hypothesis is such that:
\eqn{P(\mu_{treatment}-\mu_{control}>0) \ge 1-}\code{alpha}.
In case of a non-informative prior this coincides with the frequentist type I error.
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "stepwise")

MAPprior_cont(data = trial_data, arm = 3)


}
\author{
Katharina Hees
}
