% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_AR1_cont.R
\name{mixmodel_AR1_cont}
\alias{mixmodel_AR1_cont}
\title{Model-based analysis for continuous data adjusting for periods as a random factor with AR1 correlation structure}
\usage{
mixmodel_AR1_cont(data, arm, alpha = 0.025, ci = FALSE, ncc = TRUE, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{ci}{Boolean. Whether confidence intervals should be computed. Default=FALSE}

\item{ncc}{Boolean. Whether to include NCC data into the analysis. Default=TRUE}

\item{...}{Further arguments for simulation function}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, an indicator whether the null hypothesis was rejected or not for the investigated treatment and the fitted model
}
\description{
Performs linear mixed model regression taking into account all trial data until the arm under study leaves the trial and adjusting for periods as random factors with AR1 correlation structure
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

mixmodel_AR1_cont(data = trial_data, arm = 3, ci = TRUE)

}
\author{
Pavla Krotka
}
