% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_cal_cont.R
\name{piecewise_cal_cont}
\alias{piecewise_cal_cont}
\title{Model-based analysis for continuous data using discontinuous piecewise polynomials per calendar time}
\usage{
piecewise_cal_cont(
  data,
  arm,
  alpha = 0.025,
  unit_size = 25,
  ncc = TRUE,
  poly_degree = 3,
  check = TRUE,
  ...
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function. Must contain columns named 'treatment', 'response' and 'j'}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{unit_size}{Number of patients per calendar time unit Default=25}

\item{ncc}{Boolean. Whether to include NCC data into the analysis. Default=TRUE}

\item{poly_degree}{Degree of the piecewise polynomial. Default=3}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE}

\item{...}{Further arguments for simulation function}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, an indicator whether the null hypothesis was rejected or not for the investigated treatment and the fitted model
}
\description{
Performs linear regression taking into account all trial data until the arm under study leaves the trial and adjusting for time using discontinuous piecewise polynomials in each calendar time interval
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

piecewise_cal_cont(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
