% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_utilities.R
\name{plot_dmi}
\alias{plot_dmi}
\title{Plot Daily Dry Matter Intake Through Time}
\usage{
plot_dmi(agents)
}
\arguments{
\item{agents}{A named list of agent tibbles from a completed simulation.}
}
\value{
A ggplot object.
}
\description{
Plots daily dry matter intake through time for all agents
colored by reproductive status.
}
