% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_utilities.R
\name{plot_replicator}
\alias{plot_replicator}
\title{Plot global mean ifbfat as a function of number of agents}
\usage{
plot_replicator(output)
}
\arguments{
\item{output}{A named list returned by \code{ncc_replicator()}.}
}
\value{
A ggplot object.
}
\description{
Plots the global mean end-of-season ifbfat across replicates
as a function of number of agents, with a horizontal line indicating the
carrying capacity threshold. If the number of replicates is >= 3, a 95\%
confidence interval is plotted around the mean.
}
