% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolmodel_cont.R
\name{poolmodel_cont}
\alias{poolmodel_cont}
\title{Pooled analysis for continuous data}
\usage{
poolmodel_cont(data, arm, alpha = 0.025, check = TRUE, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function. Must contain columns named 'treatment', 'response' and 'period'}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE}

\item{...}{Further arguments for simulation function}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, an indicator whether the null hypothesis was rejected or not for the investigated treatment and the fitted model
}
\description{
Performs pooled analysis (naively pooling concurrent and non-concurrent controls without adjustment) using a linear model
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

poolmodel_cont(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
