% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{print.summary_abm}
\alias{print.summary_abm}
\title{Print an NCC agent-based model summary}
\usage{
\method{print}{summary_abm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary_abm} from \code{summary_abm}.}

\item{...}{Ignored.}
}
\description{
Print an NCC agent-based model summary
}
