% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_trend.R
\name{seasonal_trend}
\alias{seasonal_trend}
\title{Generation of a seasonal trend}
\usage{
seasonal_trend(j, lambda, n_wave, n_total)
}
\arguments{
\item{j}{Patient recruitment index.}

\item{lambda}{Strength of time trend.}

\item{n_wave}{How many cycles (waves) should the time trend have (\eqn{\psi}).}

\item{n_total}{Total sample size in the trial.}
}
\value{
Time trend for patient j.
}
\description{
Generates seasonal trend for patient j according to \eqn{f(j) = \lambda \cdot \mathrm{sin} \big( \psi \cdot 2\pi \cdot \frac{j-1}{N-1} \big)}.
}
\author{
Marta Bofill Roig, Pavla Krotka
}
