% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepmodel_bin.R
\name{sepmodel_bin}
\alias{sepmodel_bin}
\title{Separate analysis for binary data}
\usage{
sepmodel_bin(data, arm, alpha = 0.025)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on}

\item{alpha}{Type I error. Default=0.025}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Performs separate analysis (only taking into account concurrent controls) using a logistic model
}
\examples{

trial_data <- datasim_bin(n_total = 1000, num_arms = 3, d = 120,
period_blocks = 2, p0 = 0.7, OR = rep(1.4, 3), lambda = rep(0.15, 4), trend = "linear")

sepmodel_bin(data=trial_data, arm=3)

}
\author{
Pavla Krotka
}
