% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepmodel_cont.R
\name{sepmodel_cont}
\alias{sepmodel_cont}
\title{Separate analysis for continuous data}
\usage{
sepmodel_cont(data, arm, alpha = 0.025)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on}

\item{alpha}{Type I error. Default=0.025}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Performs separate analysis (only taking into account concurrent controls) using a linear model
}
\examples{

trial_data <- datasim_cont(n_total = 1000, num_arms = 3, d = 120,
period_blocks = 2, theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

sepmodel_cont(data=trial_data, arm=3)

}
\author{
Pavla Krotka
}
