% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{sim_study}
\alias{sim_study}
\title{Wrapper function performing simulation studies for a given set of scenarios}
\usage{
sim_study(nsim, scenarios, endpoint)
}
\arguments{
\item{nsim}{Number of replications}

\item{scenarios}{Data frame containing all parameters for scenarios that should be simulated}

\item{endpoint}{Endpoint indicator. "cont" for continuous endpoints, "bin" for binary endpoints}
}
\value{
List containing an indicator whether the null hypothesis was rejected or not for the investigated treatment for all models
}
\description{
Performs a simulation study for a given set of scenarios, analyzing simulated data using the fixed effect model, pooled and separate analyses, and timemachine and MAP prior approach. Performs inference for all treatment arms in the trial except for the first one
}
\author{
Pavla Krotka
}
\keyword{internal}
