% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines_cal_cont.R
\name{splines_cal_cont}
\alias{splines_cal_cont}
\title{Spline regression analysis for continuous data with knots placed according to calendar time units}
\usage{
splines_cal_cont(
  data,
  arm,
  alpha = 0.025,
  unit_size = 25,
  ncc = TRUE,
  bs_degree = 3,
  check = TRUE,
  ...
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function. Must contain columns named 'treatment', 'response' and 'j'.}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.}

\item{alpha}{Type I error rate. Default=0.025.}

\item{unit_size}{Number of patients per calendar time unit. Default=25.}

\item{ncc}{Boolean. Whether to include NCC data into the analysis. Default=TRUE.}

\item{bs_degree}{Degree of the polynomial spline. Default=3 for cubic spline.}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.}

\item{...}{Further arguments for simulation function.}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, an indicator whether the null hypothesis was rejected or not (for the investigated treatment specified in the input), the positions of the knots in terms of patient index, and the fitted model.
}
\description{
Performs linear regression taking into account all trial data until the arm under study leaves the trial and adjusting for time using regression splines with knots placed according to calendar time units.
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

splines_cal_cont(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
