% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_trend.R
\name{sw_trend}
\alias{sw_trend}
\title{Generation of stepwise trend with equal jumps between periods}
\usage{
sw_trend(cj, lambda)
}
\arguments{
\item{cj}{a}

\item{lambda}{a}
}
\value{

}
\description{

}
\references{

}
\author{

}
