% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_trend2.R
\name{sw_trend2}
\alias{sw_trend2}
\title{Generation of stepwise trend with jump sizes adapted to sample size per period}
\usage{
sw_trend2(cj, lambda, ss_period, ss_total)
}
\arguments{
\item{cj}{a}

\item{lambda}{a}

\item{ss_period}{a}

\item{ss_total}{a}
}
\value{

}
\description{

}
\references{

}
\author{

}
