% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine_cont.R
\name{timemachine_cont}
\alias{timemachine_cont}
\title{Time machine analysis for continuous data}
\usage{
timemachine_cont(
  data,
  arm,
  alpha = 0.025,
  prec_theta = 0.001,
  prec_eta = 0.001,
  tau_a = 0.1,
  tau_b = 0.01,
  prec_a = 0.001,
  prec_b = 0.001,
  bucket_size = 25,
  check = TRUE,
  ...
)
}
\arguments{
\item{data}{Trial data, e.g. result from the \code{datasim_cont()} function. Must contain columns named 'treatment', 'response' and 'period'.}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.}

\item{alpha}{Significance level. Default=0.025.}

\item{prec_theta}{Precision (\eqn{1/\sigma^2_{\theta}}) of the prior regarding the treatment effect \eqn{\theta}. I.e. \eqn{\theta \sim N(0, \sigma^2_{\theta})}. Default=0.001.}

\item{prec_eta}{Precision (\eqn{1/\sigma^2_{\eta_0}}) of the prior regarding the control mean \eqn{\eta_0} (corresponds to \code{mu0} in the \code{datasim_cont()} function). I.e. \eqn{\eta_0 \sim N(0, \sigma^2_{\eta_0})}. Default=0.001.}

\item{tau_a}{Parameter \eqn{a} of the Gamma distribution for the precision parameter \eqn{\tau} in the model for the time trend. I.e., \eqn{\tau \sim Gamma(a,b)}. Default=0.1.}

\item{tau_b}{Parameter \eqn{b} of the Gamma distribution for the precision parameter \eqn{\tau} in the model for the time trend. I.e., \eqn{\tau \sim Gamma(a,b)}. Default=0.01.}

\item{prec_a}{Parameter \eqn{a} of the Gamma distribution regarding the precision of the responses. I.e., \eqn{\sigma \sim Gamma(a,b)}. Default=0.001.}

\item{prec_b}{Parameter \eqn{b} of the Gamma distribution regarding the precision of the responses. I.e., \eqn{\sigma \sim Gamma(a,b)}. Default=0.001.}

\item{bucket_size}{Number of patients per time bucket. Default=25.}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.}

\item{...}{Further arguments for simulation function.}
}
\value{
List containing the following elements regarding the results of comparing \code{arm} to control:
\itemize{
\item \code{p-val} - posterior probability that the difference in means is less than zero
\item \code{treat_effect} - posterior mean of difference in means
\item \code{lower_ci} - lower limit of the 95\% credible interval for difference in means
\item \code{upper_ci} - upper limit of the 95\% credible interval for difference in means
\item \code{reject_h0} - indicator of whether the null hypothesis was rejected or not (\code{p_val} < \code{alpha})
}
}
\description{
This function performs analysis of continuous data using the Time Machine approach. It takes into account all data until the investigated arm leaves the trial. It is based on linear regression with treatment as a categorical variable and covariate adjustment for time via a second-order Bayesian normal dynamic linear model (separating the trial into buckets of pre-defined size).
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

timemachine_cont(data = trial_data, arm = 3)


}
\author{
Dominic Magirr, Peter Jacko
}
