% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT_all_models.R
\name{TT_all_models}
\alias{TT_all_models}
\title{Wrapper function for simulations analyzing given data with given models (treatment-treatment comparisons)}
\usage{
TT_all_models(
  data,
  arms,
  models = c("fixmodel", "indirect"),
  endpoint,
  alpha = 0.025,
  ...
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} or \code{datasim_cont()} function}

\item{arms}{Indicator of the treatment arms to be compared (vector of length 2)}

\item{models}{Vector with models that should be used for the analysis. Default=c("fixmodel", "indirect")}

\item{endpoint}{Endpoint indicator. "cont" for continuous endpoints, "bin" for binary endpoints}

\item{alpha}{Type I error. Default=0.025}

\item{...}{Further arguments for simulation function}
}
\value{
List containing an indicator whether the null hypothesis was rejected or not for the investigated treatment for all models
}
\description{
Analyzes given data using the fixed effect model and indirect comparison. Compares the indicated treatment arms.
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

TT_all_models(data = trial_data, arms = c(1,3), endpoint = "cont")


}
\author{
Pavla Krotka
}
\keyword{internal}
