% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT_indirect_cont.R
\name{TT_indirect_cont}
\alias{TT_indirect_cont}
\title{Indirect treatment-treatment comparison for continuous data}
\usage{
TT_indirect_cont(data, arms, alpha = 0.025, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function}

\item{arms}{Indicator of the treatment arms to be compared (vector of length 2)}

\item{alpha}{Type I error. Default=0.025}

\item{...}{Further arguments for simulation function}
}
\value{
List containing the p-value (one-sided), estimated difference between the treatments and an indicator whether the null hypothesis was rejected or not for the investigated comparison
}
\description{
This function performs indirect comparison of two treatments.
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

TT_indirect_cont(data = trial_data, arms = c(1,3))

}
\author{
Pavla Krotka
}
