% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT_sim_study.R
\name{TT_sim_study}
\alias{TT_sim_study}
\title{Wrapper function performing simulation studies (treatment-treatment comparisons) for a given set of scenarios}
\usage{
TT_sim_study(
  nsim,
  scenarios,
  arms,
  models = c("fixmodel", "indirect"),
  endpoint
)
}
\arguments{
\item{nsim}{Number of replications}

\item{scenarios}{Data frame containing all parameters for scenarios that should be simulated}

\item{arms}{Indicator of the treatment arms to be compared (vector of length 2)}

\item{models}{Vector with models that should be used for the analysis. Default=c("fixmodel", "indirect")}

\item{endpoint}{Endpoint indicator. "cont" for continuous endpoints, "bin" for binary endpoints}
}
\value{
Data frame with all considered scenarios and corresponding results
}
\description{
Performs a simulation study for a given set of scenarios, analyzing simulated data using the fixed effect model and indirect comparison. Compares the indicated treatment arms.
}
\author{
Pavla Krotka
}
\keyword{internal}
