% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TT_sim_study_par.R
\name{TT_sim_study_par}
\alias{TT_sim_study_par}
\title{Wrapper function performing simulation studies (treatment-treatment comparisons) for a given set of scenarios (parallelized)}
\usage{
TT_sim_study_par(
  nsim,
  scenarios,
  arms,
  models = c("fixmodel", "indirect"),
  endpoint,
  perc_cores = 0.9
)
}
\arguments{
\item{nsim}{Number of replications}

\item{scenarios}{Data frame containing all parameters for scenarios that should be simulated}

\item{arms}{Indicator of the treatment arms to be compared (vector of length 2)}

\item{models}{Vector with models that should be used for the analysis. Default=c("fixmodel", "sepmodel", "poolmodel", "timemachine", "MAPprior")}

\item{endpoint}{Endpoint indicator. "cont" for continuous endpoints, "bin" for binary endpoints}
}
\value{
Data frame with all considered scenarios and corresponding results
}
\description{
This function performs a simulation study for a given set of scenarios, analyzing simulated data using the fixed effect model and indirect comparison. It compares the indicated treatment arms.
}
\author{
Pavla Krotka
}
\keyword{internal}
