% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_models.R
\name{all_models}
\alias{all_models}
\title{Wrapper function for simulations analyzing given data with all models}
\usage{
all_models(
  data,
  arms,
  models = c("fixmodel", "sepmodel", "poolmodel", "timemachine", "MAP_rjags",
    "mixmodel"),
  endpoint,
  alpha = 0.025,
  unit_size = 250,
  opt = 1,
  prior_prec_tau = 1,
  n.samples = 1000,
  n.chains = 4,
  n.iter = 4000,
  n.adapt = 1000,
  robustify = TRUE,
  weight = 0.1,
  ci = FALSE,
  prec_delta = 0.001,
  prec_gamma = 0.001,
  tau_a = 0.1,
  tau_b = 0.01,
  prec_a = 0.001,
  prec_b = 0.001,
  bucket_size = 25,
  ...
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} or \code{datasim_cont()} function}

\item{arms}{Vector with treatment arms to perform inference on. Default - all arms except the first one}

\item{models}{Vector with models that should be used for the analysis. Default=c("fixmodel", "sepmodel", "poolmodel", "timemachine", "MAP_rjags")}

\item{endpoint}{Endpoint indicator. "cont" for continuous endpoints, "bin" for binary endpoints}

\item{alpha}{Type I error. Default=0.025}
}
\value{
List containing an indicator whether the null hypothesis was rejected or not for the investigated treatment for all models
}
\description{
Analyzes given data using the fixed effect model, pooled and separate analyses, and timemachine and MAP prior approach. Performs inference for all treatment arms in the trial except for the first one
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

all_models(data = trial_data, arms = c(2,3), endpoint = "bin")


}
\author{
Pavla Krotka
}
\keyword{internal}
