% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_functions.R
\name{calc_energy_bmr}
\alias{calc_energy_bmr}
\title{Calculate basal metabolic energy expenditure}
\usage{
calc_energy_bmr(bm)
}
\arguments{
\item{bm}{Numeric. Body mass (kg).}
}
\value{
Numeric. Basal metabolic energy expenditure (kJ/day).
}
\description{
Returns the daily basal metabolic energy expenditure from the
Chappel and Hudson (1980) basal metabolic rate (460 kJ * kg^-0.75 * day^-1),
scaled by metabolic body mass.
}
\keyword{internal}
