% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_agents.R
\name{create_agents}
\alias{create_agents}
\title{Create and initialize agent population}
\usage{
create_agents(forage_raster)
}
\arguments{
\item{forage_raster}{A \code{terra::SpatRaster}. Used to define the spatial
extent for initializing agent starting positions.}
}
\value{
A named list of tibbles, one per agent.
}
\description{
Creates a named list where each element represents an individual
sheep agent as a tibble with one row per simulation time step.
}
\keyword{internal}
