% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmodel_bin.R
\name{fixmodel_bin}
\alias{fixmodel_bin}
\title{Frequentist logistic regression model analysis for binary data adjusting for periods}
\usage{
fixmodel_bin(data, arm, alpha = 0.025, ncc = TRUE, check = TRUE, ...)
}
\arguments{
\item{data}{Trial data, e.g. result from the \code{datasim_bin()} function. Must contain columns named 'treatment', 'response' and 'period'.}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.}

\item{alpha}{Significance level (one-sided). Default=0.025.}

\item{ncc}{Boolean. Whether to include NCC data into the analysis. Default=TRUE.}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.}

\item{...}{Further arguments for simulation function.}
}
\value{
List containing the following elements regarding the results of comparing \code{arm} to control:
\itemize{
\item \code{p-val} - p-value (one-sided)
\item \code{treat_effect} - estimated treatment effect in terms of the log-odds ratio
\item \code{lower_ci} - lower limit of the (1-2*\code{alpha})*100\% confidence interval
\item \code{upper_ci} - upper limit of the (1-2*\code{alpha})*100\% confidence interval
\item \code{reject_h0} - indicator of whether the null hypothesis was rejected or not (\code{p_val} < \code{alpha})
\item \code{model} - fitted model
}
}
\description{
This function performs logistic regression taking into account all trial data until the arm under study leaves the trial and adjusting for periods as factors.
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

fixmodel_bin(data = trial_data, arm = 3)

}
\references{
On model-based time trend adjustments in platform trials with non-concurrent controls. Roig, M.B., Krotka, et al. BMC Medical Research Methodology 22.1 (2022): 1-16.
}
\author{
Pavla Krotka
}
