% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmodel_bin.R
\name{fixmodel_bin}
\alias{fixmodel_bin}
\title{Model-based analysis for binary data adjusting for periods}
\usage{
fixmodel_bin(data, arm, alpha = 0.025, ncc = TRUE, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{ncc}{Boolean. Whether to include NCC data into the analysis. Default=TRUE}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, an indicator whether the null hypothesis was rejected or not for the investigated treatment and the fitted model
}
\description{
Performs logistic regression taking into account all trial data until the arm under study leaves the trial and adjusting for periods as factors
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

fixmodel_bin(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
