% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmodel_cal_bin.R
\name{fixmodel_cal_bin}
\alias{fixmodel_cal_bin}
\title{Model-based analysis for binary data adjusting for calendar time units}
\usage{
fixmodel_cal_bin(data, arm, alpha = 0.025, unit_size = 25, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{unit_size}{Number of patients per calendar time unit Default=25}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Performs logistic regression taking into account all trial data until the arm under study leaves the trial and adjusting for calendar time units as factors
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

fixmodel_cal_bin(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
