% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixmodel_cal_cont.R
\name{fixmodel_cal_cont}
\alias{fixmodel_cal_cont}
\title{Model-based analysis for continuous data adjusting for calendar time units}
\usage{
fixmodel_cal_cont(data, arm, alpha = 0.025, unit_size = 250, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{unit_size}{Number of patients per calendar time unit Default=25}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Performs linear regression taking into account all trial data until the arm under study leaves the trial and adjusting for calendar time units as factors
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

fixmodel_cal_cont(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
