% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_param}
\alias{get_param}
\title{Get a global model parameter}
\usage{
get_param(param)
}
\arguments{
\item{param}{Character string. Name of the parameter.}
}
\description{
Get a global model parameter
}
