% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss_matrix.R
\name{get_ss_matrix}
\alias{get_ss_matrix}
\title{Sample size matrix for a platform trial with 3 arms}
\usage{
get_ss_matrix(n_total, num_arms, t_arm)
}
\arguments{
\item{n_total}{Overall sample size for the trial}

\item{num_arms}{Number of treatment arms in the trial}

\item{t_arm}{Timing of adding new arms in terms of number of patients allocated to the control arm}
}
\value{
Sample size matrix
}
\description{
Computes the sample size matrix - sample sizes per arm (rows) and per period (columns)
}
\author{
Pavla Krotka
}
\keyword{internal}
