% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_trend.R
\name{linear_trend}
\alias{linear_trend}
\title{Generation of a linear trend}
\usage{
linear_trend(j, lambda, sample_size)
}
\arguments{
\item{j}{a}

\item{lambda}{a}

\item{sample_size}{a}
}
\value{

}
\description{

}
\references{

}
\author{

}
