% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPprior_cont.R
\name{MAPprior_cont}
\alias{MAPprior_cont}
\title{MAP Prior approach analysis for continuous data}
\usage{
MAPprior_cont(
  data,
  arm,
  alpha = 0.025,
  opt = 2,
  prior_prec_tau = 4,
  n.samples = 1000,
  n.chains = 4,
  n.iter = 4000,
  n.adapt = 1000,
  robustify = TRUE,
  weight = 0.1,
  check = TRUE,
  ...
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function. Must contain columns named 'treatment', 'response' and 'period'.}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.}

\item{alpha}{Type I error rate. Default=0.025}

\item{opt}{Binary. If opt=1, all former periods are used as one source; if opt=2, periods form different sources get separately included into the final analysis. Default=2.}

\item{prior_prec_tau}{Dispersion parameter of the half normal prior, the prior for the between study heterogeneity. Default=4.}

\item{n.samples}{How many random samples will get drawn for the calculation of the posterior mean and the CIs. Default=1000.}

\item{n.chains}{Number of parallel chains for the rjags model. Default=4.}

\item{n.iter}{Number of iterations to monitor of the jags.model. Needed for coda.samples. Default=4000.}

\item{n.adapt}{Number of iterations for adaptation, an initial sampling phase during which the samplers adapt their behavior to maximize their efficiency. Needed for jags.model. Default=1000.}

\item{robustify}{Booelan.}

\item{weight}{Weight given to the non-informative component (0 < weight < 1) for the robustification of the MAP prior according to Schmidli (2014). Default=0.1.}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.}

\item{...}{Further arguments for simulation function.}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, and an indicator whether the null hypothesis was rejected or not (for the investigated treatment specified in the input).
}
\description{
This function performs analysis of continuous data using the MAP Prior approach. The method borrows data from non-concurrent controls to obtain the prior distribution for the control response in the concurrent periods.
}
\examples{

trial_data<-datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "stepwise")

MAPprior_cont(data = trial_data, arm = 3)


}
\author{
Katharina Hees
}
