% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixmodel_cont.R
\name{mixmodel_cont}
\alias{mixmodel_cont}
\title{Mixed regression model analysis for continuous data adjusting for periods as a random factor}
\usage{
mixmodel_cont(
  data,
  arm,
  alpha = 0.025,
  ci = FALSE,
  ncc = TRUE,
  check = TRUE,
  ...
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function. Must contain columns named 'treatment', 'response' and 'period'.}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1). This arm is compared to the control group.}

\item{alpha}{Type I error rate. Default=0.025.}

\item{ci}{Boolean. Whether confidence intervals should be computed. Default=FALSE.}

\item{ncc}{Boolean. Whether to include NCC data into the analysis. Default=TRUE.}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE.}

\item{...}{Further arguments for simulation function.}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, an indicator whether the null hypothesis was rejected or not (for the investigated treatment specified in the input), and the fitted model.
}
\description{
This function performs linear mixed model regression taking into account all trial data until the arm under study leaves the trial and adjusting for periods as random factors.
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

mixmodel_cont(data = trial_data, arm = 3, ci = TRUE)

}
\author{
Pavla Krotka
}
