% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{ncc_abm}
\alias{ncc_abm}
\title{Run the NCC Agent-Based Model}
\usage{
ncc_abm(forage_reference, dem, canopy, escape, verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. If \code{TRUE} (the default), progress messages are
printed as the simulation runs; if \code{FALSE}, all messages are suppressed.}
}
\description{
Run the NCC Agent-Based Model
}
