% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolmodel_bin.R
\name{poolmodel_bin}
\alias{poolmodel_bin}
\title{Pooled analysis for binary data}
\usage{
poolmodel_bin(data, arm, alpha = 0.025, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Performs pooled analysis (naively pooling concurrent and non-concurrent controls without adjustment) using a logistic model
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

poolmodel_bin(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
