% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepmodel_bin.R
\name{sepmodel_bin}
\alias{sepmodel_bin}
\title{Separate analysis for binary data}
\usage{
sepmodel_bin(data, arm, alpha = 0.025, check = TRUE, ...)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_bin()} function. Must contain columns named 'treatment', 'response' and 'period'}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{check}{Boolean. Indicates whether the input parameters should be checked by the function. Default=TRUE, unless the function is called by a simulation function, where the default is FALSE}

\item{...}{Further arguments for simulation function}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval, an indicator whether the null hypothesis was rejected or not for the investigated treatment and the fitted model
}
\description{
Performs separate analysis (only taking into account concurrent controls) using a logistic model
}
\examples{

trial_data <- datasim_bin(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
p0 = 0.7, OR = rep(1.8, 3), lambda = rep(0.15, 4), trend="stepwise")

sepmodel_bin(data = trial_data, arm = 3)

}
\author{
Pavla Krotka
}
