% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{set_param}
\alias{set_param}
\title{Set or override a global model parameter}
\usage{
set_param(param, value)
}
\arguments{
\item{param}{Character string. Name of the parameter.}

\item{value}{New value.}
}
\description{
Set or override a global model parameter
}
