% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_trend.R
\name{sw_trend}
\alias{sw_trend}
\title{Generation of stepwise trend with equal jumps between periods}
\usage{
sw_trend(cj, lambda, trend_mean, trend_var)
}
\arguments{
\item{cj}{Period indicator.}

\item{lambda}{Strength of time trend.}

\item{trend_mean}{Integer. In case of random time trends, the strength of the time trend will be generated from N(\code{trend_mean}, \code{trend_var}).}

\item{trend_var}{Integer. In case of random time trends, the strength of the time trend will be generated from N(\code{trend_mean}, \code{trend_var}).}
}
\value{
Time trend in period \eqn{c_j}.
}
\description{
This function generates a stepwise trend for a given period. No time trend is assumed in the first period.
}
\details{
The time trend is generated according to the function \eqn{f(j) = \lambda \cdot (c_j - 1)}, where \eqn{c_j} is an index of the period patient \eqn{j} was enrolled in.
}
\author{
Marta Bofill Roig, Pavla Krotka
}
