% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_trend.R
\name{sw_trend}
\alias{sw_trend}
\title{Generation of stepwise trend with equal jumps between periods}
\usage{
sw_trend(cj, lambda)
}
\arguments{
\item{cj}{Period indicator}

\item{lambda}{Strength of time trend}
}
\value{
Time trend in period cj
}
\description{
Generates stepwise trend for a given period. No time trend is assumed in the first period.
}
\author{
Marta Bofill Roig, Pavla Krotka
}
