% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_trend2.R
\name{sw_trend2}
\alias{sw_trend2}
\title{Generation of stepwise trend with jump sizes adapted to sample size per period}
\usage{
sw_trend2(cj, lambda, ss_period, ss_total)
}
\arguments{
\item{cj}{Period indicator.}

\item{lambda}{Strength of time trend.}

\item{ss_period}{Sample size in a given period.}

\item{ss_total}{Total sample size.}
}
\value{
Time trend in period \eqn{c_j}.
}
\description{
This function generates a stepwise trend for a given period. Jump sizes are adapted to the period sample size.
}
\author{
Marta Bofill Roig, Pavla Krotka
}
