% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine_cont.R
\name{timemachine_cont}
\alias{timemachine_cont}
\title{Time machine for continuous data}
\usage{
timemachine_cont(
  data,
  arm,
  alpha = 0.025,
  prec_delta = 0.001,
  prec_gamma = 0.001,
  tau_a = 0.1,
  tau_b = 0.01,
  prec_a = 0.001,
  prec_b = 0.001,
  bucket_size = 25
)
}
\arguments{
\item{data}{Simulated trial data, e.g. result from the \code{datasim_cont()} function}

\item{arm}{Indicator of the treatment arm under study to perform inference on (vector of length 1)}

\item{alpha}{Type I error. Default=0.025}

\item{prec_delta}{...}

\item{prec_gamma}{...}

\item{tau_a}{...}

\item{tau_b}{...}

\item{prec_a}{...}

\item{prec_b}{...}

\item{bucket_size}{Number of patients per time bucket. Default=25}
}
\value{
List containing the p-value (one-sided), estimated treatment effect, 95\% confidence interval and an indicator whether the null hypothesis was rejected or not for the investigated treatment
}
\description{
Time machine for continuous data
}
\examples{

trial_data <- datasim_cont(num_arms = 3, n_arm = 100, d = c(0, 100, 250),
theta = rep(0.25, 3), lambda = rep(0.15, 4), sigma = 1, trend = "linear")

timemachine_cont(data = trial_data, arm = 3)


}
\author{
Dominic Magirr, Peter Jacko
}
